pyproject.toml
[tool.poetry]
name = "my_dost"
version = "0.0.1"
description = "My-DOST is a Python based Utility platform as an Open Source project. We strive to liberate humans from mundane, repetitive tasks, giving them more time to use their intellect and creativity to solve higher-order business challenges and perform knowledge work."
authors = ["PyBots <support@pybots.ai>"]
license = "MIT"
readme = "README.md"
homepage = ""
repository = ""
documentation = "https://auto-pylot.readthedocs.io/"
keywords = ["PyBots","dost","my-dost","utility","Python","Automation", "Software BOT","Dost", "bots", "RPA"]
classifiers = [
    'Development Status :: 1 - Beta',
    'Environment :: Console',
    'Intended Audience :: Developers',
    'Intended Audience :: Education',
    'Intended Audience :: End Users/Desktop',
    'Intended Audience :: Financial and Insurance Industry',
    'Intended Audience :: Information Technology',
    'Intended Audience :: System Administrators',
    'Topic :: Software Development :: Build Tools',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Operating System :: Microsoft :: Windows',
    'Programming Language :: Python',
]
include = ["my_autopylot/*.pyd", "my_autopylot/resources/*", "my_autopylot/*.pye", "my_autopylot/__init__.py", "my_autopylot/Engine.py"]

[tool.poetry.urls]



[tool.poetry.dependencies]
python = ">=3.8.0,<=3.9.40"
SpeechRecognition = "^3.8.1"
speedtest-cli = "^2.1.3"
helium = "^3.0.8"
webdriver-manager = "^3.5.3"
PyGetWindow = "^0.0.9"
pyinspect = "^0.1.0"
pandas = "^1.4.1"
PyScreeze = "^0.1.28"
Pillow = "^9.0.1"
psutil = "^5.9.0"
openpyxl = "^3.0.9"
xlsx2html = "^0.4.0"
xlwt = "^1.3.0"
pivottablejs = "^0.9.0"
plotly = "^5.7.0"
kaleido = "0.2.1"
yagmail = "^0.15.277"
pdfplumber = "^0.6.0" 
opencv-python = "^4.5.5"
matplotlib = "^3.5.1"
imutils = "^0.5.4"
pyttsx3 = "^2.90"
sourcedefender = "^10.0.3"
xls2xlsx = "^0.1.5"
gspread = "^5.3.2"
gspread-dataframe = "^3.3.0"
oauth2client = "^4.1.3"
pendulum = "^2.1.2"
pywinauto = "^0.6.8"
pytesseract = "^0.3.9"
Cython = "^0.29.30"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
ipykernel = "^6.13.0"
pdoc3 = "^0.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
